Create The first version of py++ language
Google translator
I added two things to print 1- "Print" to print on the screen 2- "retrn" to print in console.log I also added a 'def' button which is a function inside of which you have 3 options, put in the function, a third for the condition and the third there is a slight problem that even if there is an error in the code it seems that this is the reason. I put a short message in console.log with the name As. This explains why the name was added to def